+++ src/cli-bench/Cargo.toml
@@ -18,10 +18,6 @@ version.workspace = true
 edition.workspace = true
 license.workspace = true
 
-[[bin]]
-name = "cli-bench"
-path = "src/bench.rs"
-
 [dependencies]
 common-base.workspace = true
 thiserror.workspace = true
