+++ src/cmd/Cargo.toml
@@ -35,6 +35,10 @@ path = "src/placement-center/server.rs"
 name = "cli-command"
 path = "src/cli-command/command.rs"
 
+[[bin]]
+name = "cli-bench"
+path = "src/cli-bench/bench.rs"
+
 [dependencies]
 clap.workspace = true
 common-base.workspace = true
@@ -44,6 +48,7 @@ mqtt-broker.workspace = true
 placement-center.workspace = true
 journal-server.workspace = true
 cli-command.workspace = true
+cli-bench.workspace = true
 clap-cargo.workspace = true
 protocol.workspace = true
 console-subscriber.workspace = true
